L2 gateway agent scheduler.
The patch set contains the code for the L2 gateway
scheduler that determines which L2 gateway agent
is the Monitor agent out of available agents if
there does not exist Monitor agent in the list
of active agents.
If only one L2 gateway agent is active, then
this piece of code makes that agent the Monitor
agent.
If multiple agents are active, then the logic
makes the L2 gateway agent that has the oldest
'started_at' timestamp as the Monitor agent.
It makes use of fanout cast message to send it to
all the agents. The message payload includes the host
name of the Monitor agent so that that particular
agent changes its role to the Monitor, whereas other
agents who receive this message change their role
to default transact agent.
The patch set also removes the term TRANSACT
as one state "MONITOR" is enough to decide
the role.

Co-Authored-By: Maruti Kamat <maruti.kamat@hp.com>

Change-Id: Ib06dbe320021de3653d26f267994ced52480254a