Monitor agent to monitor OVSDB servers immediately
Current logic of the L2gw agent periodically monitors
the OVSDB servers. If the plugin sets a particular
agent as the Monitor agent, the agent waits for the
next cycle to start monitoring the OVSDB servers.
The patch set makes the Monitor agent start monitoring
the OVSDB servers instantly.
The patch set also removes the periodic task and
makes use of FixedIntervalLoopingCall.

Change-Id: I2ab50d92fb3f084b46efb49b02ee17bbed3cd2ab