Retry logic for insert/delete/update MAC failures
The patch set fixes the problem described in a launchpad
bug, where any failures in inserting/updating/delete MACs
into the OVSDB server do not get retried in a later
point in time. The failure may occur when the L2 gateway
agent is down or the OVSDB server is down.

Change-Id: I79956f498c3279e1c78be812563a32c2c41e77ca
Closes-Bug: 1446805
Closes-Bug: 1451765