ovsdb server connection initiation implementation
ovsdb server initiates the connection to l2gw agent with the entries in
manager table in ovsdb hardware vtep schema.
A configurable option 'enable_manager' in .ini file is provided so that
the connection can be initiated by the ovsdb server or not based on boolean
value.
By default 'enable_manager' value is False, turn on the variable to True
to initiate the connection from ovsdb server to l2gw agent.

Closes-Bug:1466302

Change-Id: Ib668ecda381367ea9f2f3cfa45e63d0e91e48f33