Bring models in sync with migrations, add test
Bring the sqlalchemy models in sync with the alembic migrations.

Add the neutron standard model-migration sync test. The test is added
to the unit test env since l2gw does not have a functional job
yet. This requires that l2gw uses the python-db-jobs template in the
gate.

(Also remove downgrades since they are not supported.)

Closes-Bug: #1605675

Depends-On: I3a8ea02b86846e9a8a7d7f2aee9700ca952b0248

Change-Id: Ib73f554d5f0bbfbe25a0f6ad33b23e4584800e21