make tempest plugin independent of neutron-tempest-plugin
tempest-tox-plugin-sanity-check fails, and one of the many reasons is
that networking-l2gw-tempest-plugin depends on neutron-tempest-plugin.
This patch removes this dependency by adding clients to handle l2gw and
l2gw-connection REST API calls, and make the tests use these new clients.
NOTE: this patch doesn't handle the issues with the scenario tests.

Change-Id: I3d7569d4a65750b50fc5ff0eee6fb8d94d7c323a